% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A_llama}
\alias{A_llama}
\title{Llama3 Recommendations (internal)}
\format{
An object of class \code{numeric} of length 1891.
}
\usage{
A_llama
}
\value{
A numeric vector of 0 or 1.
}
\description{
Llama3 recommendations example for the illustration purpose.
}
\keyword{internal}
