\name{aihuman-package}
\alias{aihuman-package}
\alias{aihuman}
\docType{package}
\title{\packageTitle{aihuman}}
\description{\packageDescription{aihuman}}
\section{Package Content}{\packageIndices{aihuman}}
\author{\packageAuthor{aihuman}}
\section{Maintainer}{\packageMaintainer{aihuman}}
\keyword{package}
