% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_re.R
\name{CalAPCEipwRE}
\alias{CalAPCEipwRE}
\title{Compute APCE using frequentist analysis with random effects}
\usage{
CalAPCEipwRE(data, fixed, random, nAGQ = 1)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} of which columns consists of pre-treatment covariates, a binary treatment (Z), an ordinal decision (D), and an outcome variable (Y). The column names of the latter three should be specified as "Z", "D", and "Y" respectively.}

\item{fixed}{A formula for the fixed-effects part of the model to fit.}

\item{random}{A formula for the random-effects part of the model to fit.}

\item{nAGQ}{Integer scalar - the number of points per axis for evaluating the adaptive Gauss-Hermite approximation to the log-likelihood. Defaults to 1, corresponding to the Laplace approximation.}
}
\value{
An object of class \code{list} with the following elements:
\item{P.D1}{An array with dimension (k+1) by (k+2) for quantity P(D(1)=d| R=r), dimension 1 is (k+1) values of D from 0 to k, dimension 2 is (k+2) values of R from 0 to k+1.}
\item{P.D0}{An array with dimension (k+1) by (k+2) for quantity P(D(0)=d| R=r).}
\item{APCE}{An array with dimension (k+1) by (k+2) for quantity P(D(1)=d| R=r)-P(D(0)=d| R=r).}
\item{P.R}{An array with dimension (k+2) for quantity P(R=r) for r from 0 to (k+1).}
\item{alpha}{An array with estimated alpha.}
\item{delta}{An array with estimated delta.}
}
\description{
Estimate propensity score and use Hajek estimator to compute APCE. See S7 for more details.
}
\examples{
\donttest{
data(synth)
data(hearingdate_synth)
synth$CourtEvent_HearingDate <- hearingdate_synth
freq_apce_re <- CalAPCEipwRE(synth, 
                   fixed = "Y ~ Sex + White + Age + 
                   CurrentViolentOffense + PendingChargeAtTimeOfOffense + 
                   PriorMisdemeanorConviction + PriorFelonyConviction + 
                   PriorViolentConviction + D",
                   random = "~ 1|CourtEvent_HearingDate")
}

}
\references{
Imai, K., Jiang, Z., Greiner, D.J., Halen, R., and Shin, S. (2023).
"Experimental evaluation of algorithm-assisted human decision-making: 
application to pretrial public safety assessment." 
Journal of the Royal Statistical Society: Series A.
<DOI:10.1093/jrsssa/qnad010>.
}
