% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ability_compute_nuisance_functions.R
\name{crossfit}
\alias{crossfit}
\alias{gbm}
\title{Crossfitting for nuisance functions}
\usage{
crossfit(data, include_for_fit, form, ...)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix}  to fit on.}

\item{include_for_fit}{Boolean vector for whether or not a unit should be includeded in fitting (e.g. treated/control).}

\item{form}{Formula for outcome regression/propensity score models.}

\item{...}{Additional arguments to be passed to \code{gbm} function.}
}
\value{
A vector of predicted values
}
\description{
Implement crossfitting with boosting methods and get predicted values for outcome/decision regression or propensity score models
}
