% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_decision.R
\name{g_legend}
\alias{g_legend}
\title{Pulling ggplot legend}
\usage{
g_legend(p)
}
\arguments{
\item{p}{A \code{ggplot} of which legend should be pulled.}
}
\value{
A ggplot legend.
}
\description{
Pulling ggplot legend
}
