% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hearingdate_synth}
\alias{hearingdate_synth}
\title{Synthetic court event hearing date}
\format{
A date variable.
}
\usage{
hearingdate_synth
}
\description{
A synthetic court event hearing date
}
\keyword{datasets}
