% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nca_follow_policy}
\alias{nca_follow_policy}
\title{NCA follow policy (internal; increasing monotonicity)}
\format{
A data frame with 41 rows and 6 variables:
\describe{
  \item{FTAScore}{FTA score}
  \item{NCAScore}{NCA score}
  \item{NVCAFlag}{NVCA flag}
  \item{wts}{Weight for each observation}
  \item{n}{Number of such cases}
  \item{policy}{Optimal policy}
}
}
\usage{
nca_follow_policy
}
\description{
Optimal policy whether to follow PSA recommendation for NCA as an outcome
and policy class with an increasing monotonicity constraint.
Used for the replication in the vignette.
}
\keyword{internal}
