% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuis_func}
\alias{nuis_func}
\title{Nuisance functions (internal)}
\format{
An object of class \code{nuisance_functions} of length 2.
}
\usage{
nuis_func
}
\value{
A list with the following components:
\describe{
  \item{\code{z_models}}{A \code{data.frame} with the following columns:
    \describe{
      \item{\code{idx}}{Index of observation.}
      \item{\code{d_pred}}{Predicted probability of decision.}
      \item{\code{y_pred}}{Predicted probability of outcome.}
      \item{\code{Z}}{Treatment group.}
    }
  }
  \item{\code{pscore}}{A vector of predicted propensity scores.}
}
}
\description{
Nuisance functions generated with `compute_nuisance_functions` for the illustration purpose.
}
\keyword{internal}
