% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt.R
\name{PlotSpilloverCRT}
\alias{PlotSpilloverCRT}
\title{Plot conditional randomization test}
\usage{
PlotSpilloverCRT(res)
}
\arguments{
\item{res}{A \code{list} generated with \code{SpilloverCRT}.}
}
\value{
A ggplot
}
\description{
See Figure S8 for example.
}
\examples{
\dontrun{
data(synth)
data(hearingdate_synth)
crt <- SpilloverCRT(D = synth$D, Z = synth$Z, CourtEvent_HearingDate = hearingdate_synth)
PlotSpilloverCRT(crt)
}

}
