% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crt.R
\name{PlotSpilloverCRTpower}
\alias{PlotSpilloverCRTpower}
\title{Plot power analysis of conditional randomization test}
\usage{
PlotSpilloverCRTpower(res)
}
\arguments{
\item{res}{A \code{data.frame} generated with \code{SpilloverCRTpower}.}
}
\value{
A ggplot
}
\description{
See Figure S9 for example.
}
\examples{
\dontrun{
data(synth)
data(hearingdate_synth)
crt_power <- SpilloverCRTpower(D = synth$D, Z = synth$Z, CourtEvent_HearingDate = hearingdate_synth)
PlotSpilloverCRTpower(crt_power)
}

}
