% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psa_synth}
\alias{psa_synth}
\title{Synthetic PSA data}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
  \item{Z}{binary treatment}
  \item{D}{ordinal decision}
  \item{FTAScore}{FTA score}
  \item{NCAScore}{NCA score}
  \item{NVCAFlag}{NVCA flag}
  \item{DMF}{DMF recommendation}
}
}
\usage{
psa_synth
}
\description{
A synthetic dataset containing a binary treatment (Z), ordinal decision (D), three PSA variables (FTAScore, NCAScore, and NVCAFlag), and DMF recommendation.
}
\keyword{datasets}
