% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synth}
\alias{synth}
\title{Synthetic data}
\format{
A data frame with 1000 rows and 11 variables:
\describe{
  \item{Z}{binary treatment}
  \item{D}{ordinal decision}
  \item{Y}{outcome}
  \item{Sex}{male or female}
  \item{White}{white or non-white}
  \item{Age}{age}
  \item{CurrentViolentOffense}{binary variable for current violent offense}
  \item{PendingChargeAtTimeOfOffense}{binary variable for pending charge (felony, misdemeanor, or both) at the time of offense}
  \item{PriorMisdemeanorConviction}{binary variable for prior conviction of misdemeanor}
  \item{PriorFelonyConviction}{binary variable for prior conviction of felony}
  \item{PriorViolentConviction}{four-level ordinal variable for prior violent conviction}
}
}
\usage{
synth
}
\description{
A synthetic dataset containing pre-treatment covariates, a binary treatment (Z), an ordinal decision (D), 
and an outcome variable (Y).
}
\keyword{datasets}
