% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist.R
\name{TestMonotonicity}
\alias{TestMonotonicity}
\title{Test monotonicity}
\usage{
TestMonotonicity(data)
}
\arguments{
\item{data}{A \code{data.frame} or \code{matrix} of which columns consists of pre-treatment covariates, a binary treatment (Z), an ordinal decision (D), and an outcome variable (Y). The column names of the latter three should be specified as "Z", "D", and "Y" respectively.}
}
\value{
Message indicating whether the monotonicity assumption holds.
}
\description{
Test monotonicity using frequentist analysis
}
\examples{
data(synth)
TestMonotonicity(synth)

}
