% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ability_compute_agreement.R
\name{vis_agreement}
\alias{vis_agreement}
\title{Visualize Agreement (internal)}
\usage{
vis_agreement(
  df,
  p.title = NULL,
  p.lb = -0.2,
  p.ub = 0.2,
  y.lab = "Impact of PSA"
)
}
\arguments{
\item{df}{A data frame generated by \code{compute_stats_agreement} and reshaped.}

\item{p.title}{A title for the plot (character). Default NULL.}

\item{p.lb}{A lower bound for the y-axis (numeric). Default -0.2.}

\item{p.ub}{An upper bound for the y-axis (numeric). Default 0.2.}

\item{y.lab}{A label for the y-axis (character). Default "Impact of PSA".}
}
\value{
A ggplot object.
}
\description{
Internal function to visualize the agreement between human decisions and AI recommendations using a difference-in-means estimator of an indicator \eqn{1\{D_i = A_i\}}.
}
\keyword{internal}
