% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuis_func}
\alias{nuis_func}
\title{Nuisance functions (internal)}
\format{
An object of class \code{list} of length 2.
}
\usage{
nuis_func
}
\value{
A list with following components:
\item{z_models}{A \code{data.frame} with following columns:
\itemize{
\item{idx}{index of observation}
\item{d_pred}{predicted probability of decision}
\item{y_pred}{predicted probability of outcome}
\item{Z}{treatment group}
}}
\item{pscore}{A vector of predicted propensity scores}
}
\description{
Nuisance functions generated with `compute_nuisance_functions` for the illustration purpose.
}
\keyword{internal}
